Save or cancel renaming on focus remove.

+++ src-built-in/components/appLauncher2/src/components/FoldersList.jsx
@@ -21,7 +21,8 @@ export default class FoldersList extends React.Component {
 		this.changeFolderName = this.changeFolderName.bind(this)
 		this.onFoldersListUpdate = this.onFoldersListUpdate.bind(this)
 		this.keyPressed = this.keyPressed.bind(this)
-		this.deleteFolder = this.deleteFolder.bind(this);
+		this.deleteFolder = this.deleteFolder.bind(this)
+		this.onFocusRemove = this.onFocusRemove.bind(this)
 	}
 
 	onAppDrop(event, folder) {
@@ -52,6 +53,26 @@ export default class FoldersList extends React.Component {
 		})
 	}
 
+	onFocusRemove(event) {
+		// We don't want to hide the input if user clicked on it
+		// We only hide when the click is anywhere else in the document
+		if(event.target.id === 'rename') {
+			return
+		}
+		// If focus removed and nothing was type, then just hide
+		// and consider it a rename cancel
+		if(!this.state.folderNameInput) {
+			// Cancel rename
+			this.setState({
+				renamingFolder: null
+			})
+			this.removeClickListener()
+			return
+		}
+		//Finally, all good and so we can rename the folder
+		this.attempRename()
+	}
+
 	componentWillMount() {
 		getStore().addListener({ field: 'appFolders.list' }, this.onFoldersListUpdate)
 	}
@@ -65,7 +86,8 @@ export default class FoldersList extends React.Component {
 		e.stopPropagation();
 		this.setState({
 			renamingFolder: name
-		});
+		})
+		this.addClickListener()
 	}
 
 	changeFolderName(e) {
@@ -77,40 +99,57 @@ export default class FoldersList extends React.Component {
 	deleteFolder(name, e) {
 		e.stopPropagation();
 		e.preventDefault();
-		storeActions.deleteFolder(name);
+		// Do not attemp to delete if user is renaming a folder
+		!this.state.renamingFolder && storeActions.deleteFolder(name)
 	}
 
 	keyPressed(e) {
 		if (e.key === "Enter") {
-			const input = this.state.folderNameInput.trim()
-			const oldName = this.state.renamingFolder, newName = input
-			// Check user input to make sure its at least 1 character
-			// made of string, number or both
-			if (!/^([a-zA-Z0-9\s]{1,})$/.test(input)) {
-				// Do not rename
-				console.warn('A valid folder name is required. /^([a-zA-Z0-9\s]{1,})$/')
-				return
-			}
-			this.setState({
-				folderNameInput: "",
-				renamingFolder: null
-			}, () => {
-				storeActions.renameFolder(oldName, newName);
-			});
+			this.attempRename()
 		}
+	}
+
+	addClickListener() {
+		document.addEventListener('click', this.onFocusRemove)
+	}
+
+	removeClickListener() {
+		document.removeEventListener('click', this.onFocusRemove)
+	}
+	/**
+	 * To be called when user press Enter or when focus is removed
+	 */
+	attempRename() {
+		const input = this.state.folderNameInput.trim()
+		const oldName = this.state.renamingFolder, newName = input
+		// Check user input to make sure its at least 1 character
+		// made of string, number or both
+		if (!/^([a-zA-Z0-9\s]{1,})$/.test(input)) {
+			// Do not rename
+			console.warn('A valid folder name is required. /^([a-zA-Z0-9\s]{1,})$/')
+			return
+		}
+		this.setState({
+			folderNameInput: "",
+			renamingFolder: null
+		}, () => {
+			storeActions.renameFolder(oldName, newName)
+			// No need for the click listener any more
+			this.removeClickListener()
+		})
 	}
 
 	renderFoldersList() {
 		const folders = storeActions.getFolders()
 		return this.state.foldersList.map((folderName, index) => {
-			const folder = folders[folderName];
+			const folder = folders[folderName]
 			let className = 'complex-menu-section-toggle'
 			if (this.state.activeFolder === folderName) {
 				className += ' active-section-toggle'
 			}
 
 			let nameField = folder.icon === 'ff-folder' && this.state.renamingFolder === folderName ?
-				<input value={this.state.folderNameInput}
+				<input id="rename" value={this.state.folderNameInput}
 					onChange={this.changeFolderName}
 					onKeyPress={this.keyPressed} autoFocus /> : folderName
